Fix compatibility with newest main
