+++ src/main/java/Sessions/CardShuffler.java
@@ -37,6 +37,26 @@ public abstract class CardShuffler implements Observer{
     }
 
     /**
+     * A method for testing that adds a throwaway flashcard to this shuffler's flashcard to flashcard data map.
+     * @return The flashcard that was added
+     */
+    public Flashcard addToFlashcardData() {
+        Flashcard newFlashcard = new Flashcard(new Flashcard.Front("Hi", null), "Hi");
+        this.flashcardToData.put(newFlashcard, new FlashcardData(0));
+        return newFlashcard;
+    }
+
+    /**
+     * A method for testing that removes the first flashcard from this shuffler's flashcard to flashcard data map.
+     * @return The Flashcard that was removed
+     */
+    public Flashcard removeFromFlashcardData() {
+        Flashcard removedFlashcard = this.flashcardToData.keySet().stream().toList().get(0);
+        this.flashcardToData.remove(removedFlashcard);
+        return removedFlashcard;
+    }
+
+    /**
      * getter method for the deckCopy
      * @return List<Flashcard>
      */
