+++ src/main/java/Sessions/SmartShuffle.java
@@ -97,26 +97,6 @@ public class SmartShuffle extends CardShuffler implements UpdatingShuffler {
         return this.lastFlashcardShown;
     }
 
-    /**
-     * A method for testing that adds a throwaway flashcard to this shuffler's flashcard to flashcard data map.
-     * @return The flashcard that was added
-     */
-    public Flashcard addToFlashcardData() {
-        Flashcard newFlashcard = new Flashcard(new Flashcard.Front("Hi", null), "Hi");
-        this.flashcardToData.put(newFlashcard, new FlashcardData(0));
-        return newFlashcard;
-    }
-
-    /**
-     * A method for testing that removes the first flashcard from this shuffler's flashcard to flashcard data map.
-     * @return The Flashcard that was removed
-     */
-    public Flashcard removeFromFlashcardData() {
-        Flashcard removedFlashcard = this.flashcardToData.keySet().stream().toList().get(0);
-        this.flashcardToData.remove(removedFlashcard);
-        return removedFlashcard;
-    }
-
     public Map<Flashcard, FlashcardData> getFlashcardToData(){
         return this.flashcardToData;
     }
